+++ jpro-routing/core/src/main/java/module-info.java
@@ -4,13 +4,13 @@ module one.jpro.platform.routing.core {
 
     requires transitive de.sandec.jnodes;
     requires transitive one.jpro.jmemorybuddy;
+    requires transitive one.jpro.platform.utils;
     requires transitive jpro.webapi;
     requires transitive simplefx.core;
     requires transitive simplefx.utility;
     requires transitive simplefx.wrapping;
     requires transitive simplefx.extended;
     requires transitive scala.library;
-    requires transitive one.jpro.platform.internal.openlink;
 
     exports one.jpro.platform.routing;
     exports one.jpro.platform.routing.crawl;